Moves lib files to shared
