+++ src/Dialog/Dialog.js
@@ -1,6 +1,6 @@
 import AriaComponent from '../AriaComponent';
-import getElementPair from '../lib/getElementPair';
-import interactiveChildren from '../lib/interactiveChildren';
+import getElementPair from '../shared/getElementPair';
+import interactiveChildren from '../shared/interactiveChildren';
 
 /**
  * Class to set up an interactive Dialog element.
