+++ src/shared/getFirstAndLastItems.test.js
