+++ src/shared/interactiveChildren.test.js
